convert: fix bugs
- use `uint16` which is unsigned so we can not have a negative integer
  to convert
- given the rule of max 3 consecutive symbols, we can not represent a
  number greater than 3999
- `uint16` is a 16 bit integer that can store max value of 65535

+++ 15_property_testing/roman_numerals.go
@@ -2,7 +2,7 @@ package romans
 
 import "strings"
 
-func ConvertToRomans(n int) string {
+func ConvertToRomans(n uint16) string {
 	var results strings.Builder
 
 	for _, numeral := range allRomanNumerals {
@@ -15,7 +15,7 @@ func ConvertToRomans(n int) string {
 	return results.String()
 }
 
-func ConvertToArabic(roman string) (total int) {
+func ConvertToArabic(roman string) (total uint16) {
 	for _, symbols := range windowedRoman(roman).Symbols() {
 		total += allRomanNumerals.ValueOf(symbols...)
 	}
@@ -23,13 +23,13 @@ func ConvertToArabic(roman string) (total int) {
 }
 
 type romanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
 type romanNumerals []romanNumeral
 
-func (r romanNumerals) ValueOf(symbols ...byte) int {
+func (r romanNumerals) ValueOf(symbols ...byte) uint16 {
 	symbol := string(symbols)
 	for _, s := range r {
 		if s.Symbol == symbol {
